リファクタリング: コンポーネントをファイルに切り出し
