+++ frontend/app/src/App.js
@@ -1,167 +1,10 @@
 import { useState, useCallback, useEffect } from "react";
-import Dropzone from "react-dropzone";
 
-import { detectFace, recognizeFace } from "./lib/api";
 import "./App.css";
-
-function loadFile(file) {
-  return new Promise((resolve, reject) => {
-    const fileReader = new FileReader();
-    fileReader.addEventListener("load", () => {
-      resolve(fileReader.result);
-    }, false);
-    fileReader.readAsDataURL(file);
-  });
-}
-
-function getImageSize(url) {
-  return new Promise((resolve, reject) => {
-    var image = new Image() ;
-    image.addEventListener("load", () => {
-      resolve({
-        width: image.naturalWidth,
-        height: image.naturalHeight,
-      });
-    }, false);
-    image.src = url;
-  });
-}
-
-function ImageDropzone({ onImageDrop = () => {}, acceptableTypes = [], children = null }) {
-  const onDrop = useCallback(async (acceptedFiles) => {
-    if ( acceptedFiles.length < 0 ) return;
-    if ( acceptableTypes.indexOf(acceptedFiles[0].type) < 0 ) return;
-    const file = acceptedFiles[0];
-    const dataUrl = await loadFile(file);
-    const { width, height } = await getImageSize(dataUrl);
-    onImageDrop({ file, dataUrl, width, height });
-  }, [onImageDrop]);
-
-  return (
-    <Dropzone onDrop={onDrop}>
-      {({getRootProps, getInputProps}) => (
-        <div {...getRootProps()}>
-          <input {...getInputProps()} />
-          {children}
-        </div>
-      )}
-    </Dropzone>
-  );
-}
-
-const colorMap = {
-  "M": "#0000CC",
-  "F": "#CC0000",
-};
-
-function BoundingBoxRect({ boundingBox, ...props }) {
-  const { x1, y1, x2, y2 } = boundingBox;
-  return (
-    <rect
-        x={x1}
-        y={y1}
-        width={x2 - x1}
-        height={y2 - y1}
-        {...props} />
-  );
-}
-
-/*
-function offsetBoundingBox(boundingBox, offset) {
-  const { x1, y1, x2, y2 } = boundingBox;
-  return {
-    x1: x1 - offset,
-    y1: y1 - offset,
-    x2: x2 + offset,
-    y2: y2 + offset,
-  };
-}
-*/
-
-function BoundingBox({ face, selected, onClick }) {
-  const color = colorMap[face.sex];
-  const alpha = (selected ? "CC" : "99");
-  return (
-    <>
-      <BoundingBoxRect
-          boundingBox={face.boundingBox}
-          stroke={"#FFFFFF" + alpha}
-          strokeWidth={selected ? 5 : 3}
-          fill="none" />
-      <BoundingBoxRect
-          style={{cursor: "pointer"}}
-          boundingBox={face.boundingBox}
-          stroke={color + alpha}
-          strokeWidth={selected ? 3 : 1}
-          fill="#FFFFFF00"
-          onClick={onClick} />
-    </>
-  );
-};
-
-function Face({ face, selected, onClick }) {
-  return (
-    <BoundingBox
-        face={face}
-        selected={selected}
-        onClick={onClick} />
-  );
-}
-
-function Faces({ faces, selectedFaceIndex, onClick }) {
-  return (
-    <>
-      {faces.map((face, index) => (
-        <Face
-            key={index}
-            face={face}
-            selected={index === selectedFaceIndex}
-            onClick={() => onClick({ index, face })} />
-      ))}
-    </>
-  );
-}
-
-function DetectionResultImage({ imageUrl, imageWidth, imageHeight, faces, selectedFaceIndex, onFaceClick }) {
-  return (
-    <svg
-        xmlns="http://www.w3.org/2000/svg"
-        version="1.1"
-        viewBox={`0 0 ${imageWidth} ${imageHeight}`}
-        width={imageWidth}
-        height={imageHeight}>
-      <image
-          x={0}
-          y={0}
-          width={imageWidth}
-          height={imageHeight}
-          href={imageUrl} />
-      <Faces
-          faces={faces}
-          selectedFaceIndex={selectedFaceIndex}
-          onClick={onFaceClick} />
-    </svg>
-  );
-}
-
-function FaceCroppedImage({ imageUrl, imageWidth, imageHeight, face, faceWidth = 200, faceHeight = 200 }) {
-  const { x1, y1, x2, y2 } = face.boundingBox;
-  return (
-    <svg
-        xmlns="http://www.w3.org/2000/svg"
-        version="1.1"
-        viewBox={`${x1} ${y1} ${x2 - x1} ${y2 - y1}`}
-        width={faceWidth}
-        height={faceHeight}>
-      <image
-          x={0}
-          y={0}
-          width={imageWidth}
-          height={imageHeight}
-          href={imageUrl} />
-    </svg>
-  );
-}
+import { detectFace, recognizeFace } from "./lib/api";
+import FaceCroppedImage from "./component/FaceCroppedImage";
+import FaceDetectionImage from "./component/FaceDetectionImage";
+import ImageDropzone from "./component/ImageDropzone";
 
 function chooseFaceIndex(faces) {
   if ( faces.length <= 0 ) return null;
@@ -267,7 +110,7 @@ export default function App() {
         <>
           <h1>Result</h1>
           <div>
-            <DetectionResultImage
+            <FaceDetectionImage
                 imageUrl={image.dataUrl}
                 imageWidth={image.width}
                 imageHeight={image.height}
@@ -295,9 +138,9 @@ export default function App() {
                       </td>
                       <td>
                         <img
-                          src={actor.fanza.faceImage.url}
-                          width={125}
-                          height={125} />
+                            src={actor.fanza.faceImage.url}
+                            width={125}
+                            height={125} />
                       </td>
                       <td>{actor.similarity}</td>
                     </tr>
